requirements.txt
fastapi>=0.118.3
uvicorn>=0.38.0
pydantic>=2.12.0
pillow>=12.0.0
reportlab==4.2.2
python-multipart==0.0.9
openpyxl>=3.1.5
pandas>=2.2.3